Switch to Laminas
Laminas was installed by the composer.json but it was decided to have
the CMS contain the packages. But everything is ready to externalize
them again.
The CMS previously mainly used the autloader and Zend_Mail
functionality. This is still the case and changes include the migration
to the new class names and their subtleties.
There's a new function to acts like a factory for mail transport objects
based on the arguments. See zendmail.php to see how to call it.