Require myfunctions whenever config.php is included
As the config.php now uses a class the class loading facility has to be
ready. For most pages this is automatically done when including the
connect.inc.php as this is sourced by most pages, but not all. The
upload facilities only source the config, which then breaks. Now these
first requie the myfunction.inc.php before including the config and in
addition, to be fool proof, the config itself now requires the
myfunction.inc.php and is now self-contained.
No duplicate declaration happens, as require_once is used everywhere for
the myfunction include file.