Revert automatic latin1 -> utf-8 conversions
As the CMS is (almost) purely ISO-8859-1 (latin1) based, in the inital
refactorings an automatic conversion occured when editing files as the
project base configuration was working with utf-8. This caused some
troubles especially when importing data.
The files are now reverted to the old character encoding and the
project's defeault charset has been set to latin1.