Fix char encoding issue with get param in mailing
Previously the subject read with prompt() was not correctly in
javascript when sending it as a GET param to the backend. This resulted
in an empty string.

The subject is now correctly urlencoded using encodeURIComponent on the
javascript side, which sends an enoded utf-8 string. On the PHP side the
utf-8 string is then encoded using mb_encode_numericentity() and then
converted to latin1 using utf8_decode().