Fix fatal error with a non-functional change
Troubleshooting showed that this short-circuit first check does not
result in a correct login. Instead of roubleshooting this it was decided
to have the first login fail as it never succeeded anyway and rely on
the previous functionality.
To not allow bogey login it was decided to take a column that would
never contain a valid email, e.g. the hash column.