Refactored whole navigation
The principle is still the same, but code has been outsourced to
helper functions and loops have been changed from for to foreach.
The unordered list are now handled upfront by investigating the level only
and then opened/closed based on level change. This makes
most of these extra variables like subsub, wassub etc. obsolete.

The CSS code has been updated for  #navLeft  and and  #navLeftSubTop . The
respective database updates are added to the  r/runonce.php  for migration
of existent CMS instances.